acmicpc/src/test/java/problem/step03/Q2439Test.java
package problem.step03;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;

class Q2439Test {

    @ParameterizedTest(name = "{index} [{arguments}]")
    @MethodSource
    void test(int x, String[] expected) throws Exception {
        assertThat(Q2439.solution(x)).isEqualTo(expected);
    }

    private static Stream<Arguments> test() {
        return Stream.of(
                Arguments.of(5, new String[]{
                        "    *",
                        "   **",
                        "  ***",
                        " ****",
                        "*****",
                })
        );
    }
}